# core/openapi.yaml
openapi: 3.0.3
info:
  title: Pen CMS API Specification
  description: |
    The contract-first API specification for Pen CMS, a Git-native flat-file CMS.

    Humans and agents share one capability vocabulary (string lists). `write:*` is
    create+update (never split create:/update:). `delete:*` is separate. `read` is
    monolithic. `publish:content` is content status (PATCH approve/publish) and is
    not host `publish` (deploy + Deploy Grant). Generic PUT of a post/page does not
    sniff status — that stays under write:posts / write:pages.

    Legacy agent scope `write` expands one-way to every write:*, every delete:*, and
    publish:content. It never implies host publish, users:manage, or manage:sites.

    Human admin routes under /api authorize from cookie `pen_jwt` or Bearer JWT →
    user YAML. Forgeable `pen_role` / `pen_user_id` cookies are not a root of trust.
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /comments:
    get:
      summary: List visible comments for a post
      description: >
        Unauthenticated public thread. Returns `visibility: visible` comment
        files beside the post (`content/sites/{site}/{post}/comments/c-*.md`).
        Top-level comments are oldest-first by `received_at` then slug; each
        comment's replies follow immediately after it, oldest-first among
        siblings. Pending and hidden are never included. Site binding: Host domain match wins; otherwise
        `X-Pen-Site-Id` (live PHP InternalAPIClient), query `site`, then
        `pen_site_id` cookie. Client body `site_id` is ignored. Unknown or missing posts return an empty
        `comments` array (HTTP 200). Missing `post_slug` is 400. Not an MCP
        tool.
      parameters:
        - name: post_slug
          in: query
          required: true
          schema:
            type: string
          description: Slug of the parent post or page.
      responses:
        '200':
          description: Visible comments (possibly empty).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicCommentList'
        '400':
          description: Missing post_slug.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /mcp/comments:
    get:
      operationId: list_comments
      summary: MCP list comments beside a post
      description: >
        Agent list of comment files beside the post. JWT `site_id` is
        authoritative (`X-Pen-Site-Id` cannot retarget). Omit `visibility`
        for all states (admin payload including pending); set it to filter.
        Oldest-first by `received_at` then slug. Unknown post → empty
        `comments` array. Requires `read`. Not the public GET `/comments`
        (visible-only, no `visibility` field).
      security:
        - ApiKeyAuth: []
      parameters:
        - name: post_slug
          in: query
          required: true
          schema:
            type: string
        - name: visibility
          in: query
          required: false
          schema:
            type: string
            enum: [pending, visible, hidden]
      responses:
        '200':
          description: Comment files for the post (possibly empty).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCommentList'
        '400':
          description: Missing post_slug or invalid visibility.
        '401':
          description: Not authenticated.
        '403':
          description: Missing read on this site.
    post:
      operationId: create_comment
      summary: MCP create a visible agent comment
      description: >
        Write `comments/c-*.md` beside the post with `visibility: visible`,
        `author_kind: agent`, `source_type: mcp`. Display `author_name` and
        `agent_key_name` are the JWT key name. Unknown `post_slug` is 400.
        Do not use `create_post` for a thread reply. Requires `write:posts`.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComment'
      responses:
        '200':
          description: Created comment.
          content:
            application/json:
              schema:
                type: object
                required:
                  - comment
                properties:
                  comment:
                    $ref: '#/components/schemas/AdminComment'
        '400':
          description: Unknown post_slug or invalid in_reply_to.
        '401':
          description: Not authenticated.
        '403':
          description: Missing write:posts on this site.
...
  /feedback:
    post:
      summary: Ingest public feedback
      description: >
        Unauthenticated contact or comment ingest. Site binding: Host domain
        match wins; otherwise `X-Pen-Site-Id` (PHP InternalAPIClient), query
        `site`, `source_url` `?site=` (live `/blog/?site=` preview), then the
        `pen_site_id` cookie. Body `site_id` is ignored (relay contract).
        `kind=contact` writes a canonical stub page at
        `content/sites/{site_id}/fb-{timestamp}-{slug}/index.md` via write_page.
        `kind=comment` requires `parent_slug` of an existing post/page (not an
        `fb-*` stub) and writes `content/sites/{site_id}/{post}/comments/c-*.md`
        with `visibility: pending`. Unknown `parent_slug` is 400. Rate limit:
        5 submissions per IP per 5 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackIngest'
            example:
              name: Test
              message: Hello from curl
              kind: contact
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/FeedbackIngest'
      responses:
        '200':
          description: Submission stored (contact stub or pending comment file).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackReceived'
        '400':
          description: >
            Validation error (missing message, oversize field, invalid kind,
            missing parent_slug for comments, unknown post_slug).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '429':
          description: Too many submissions from this IP (5 per 5 minutes).
  /feedback/sync:
    post:
      summary: Sync remote feedback from the public relay
      description: >
        Drain unacked items from the site's feedback relay (`GET /fetch` then
        `POST /ack`) via `write_ingest_item` (`source_type: relay`). Contact
        items become `fb-*` stubs; comments become `comments/c-*.md` beside
        the parent post. Invalid comment parents are skipped (logged) and still
        acked so the queue does not jam. Human session or agent scope `write`.
        Site is the agent JWT `site_id` or the human active site — not Host.
        Missing keys return HTTP 200 `{written: 0, reason: no_relay_configured}`.
        A down relay returns `{written: 0, reason: relay_unreachable}`. Never
        500 for those cases. Does not mint tokens.
      security:
        - CookieJwtAuth: []
        - BearerJwt: []
      responses:
        '200':
          description: Sync result (including empty / unreachable).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackSyncResult'
        '401':
          description: Not authenticated.
        '403':
          description: Authenticated actor lacks scope write.
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Pen-API-Key
    CookieJwtAuth:
      type: apiKey
      in: cookie
      name: pen_jwt
      description: Human session JWT. Authorize from this cookie → YAML. Never trust pen_role / pen_user_id.
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human or agent JWT. Agent tokens use type=agent + scopes + site_id; they do not inherit the sponsor admin role.
...
  schemas:
...
    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
    FeedbackIngest:
      type: object
      required:
        - message
      properties:
        name:
          type: string
          maxLength: 120
          description: Optional submitter handle. Stored as frontmatter submitter, not page name.
        message:
          type: string
          minLength: 1
          maxLength: 8000
          description: Raw message body written to the stub page.
        email:
          type: string
          description: Optional, not rendered publicly.
        source_url:
          type: string
          description: Optional URL of the page the visitor submitted from.
        kind:
          type: string
          enum: [contact, comment]
          default: contact
        parent_slug:
          type: string
          description: >
            Post slug for comments; sanitized. Required when kind=comment.
            Unknown post (or an fb-* stub) returns 400. Ignored for contact.
        site_id:
          type: string
          description: >
            Ignored. Write target is Host when it matches a site domain,
            else X-Pen-Site-Id, else query `site`, else `source_url` `?site=`,
            else `pen_site_id` cookie.
    FeedbackReceived:
      type: object
      required:
        - status
        - slug
      properties:
        status:
          type: string
          enum: [received]
        slug:
          type: string
          description: >
            Contact stub id (`fb-…`) or comment id (`c-…`), e.g.
            fb-20260819120000-hello-from-curl or c-20260820t153000z-nice-frosting.
    FeedbackSyncResult:
      type: object
      required:
        - written
      properties:
        written:
          type: integer
          minimum: 0
          description: Number of contact stubs plus comment files written this drain.
        slugs:
          type: array
          items:
            type: string
          description: fb-* and/or c-* ids written this drain.
        reason:
          type: string
          enum: [no_relay_configured, relay_unreachable]
          description: >
            Present when nothing could be drained, or when ack failed after
            writes. Omitted on a successful drain (including empty queue).
    PublicComment:
      type: object
      required:
        - slug
        - author_name
        - author_kind
        - body
        - received_at
      properties:
        slug:
          type: string
          description: Comment file id (`c-…`).
        author_name:
          type: string
          description: Display name only. No commenter accounts or email.
        author_kind:
          type: string
          enum: [public, agent, human]
        body:
          type: string
          description: Comment markdown/plain body.
        in_reply_to:
          type: string
          nullable: true
          description: Parent comment slug; one hop only. Null for top-level.
        received_at:
          type: string
          description: UTC timestamp when the comment was stored.
    PublicCommentList:
      type: object
      required:
        - post_slug
        - comments
      properties:
        post_slug:
          type: string
        comments:
          type: array
          items:
            $ref: '#/components/schemas/PublicComment'
    AdminComment:
      type: object
      required:
        - slug
        - author_name
        - author_kind
        - body
        - received_at
        - visibility
      properties:
        slug:
          type: string
          description: Comment file id (`c-…`).
        author_name:
          type: string
          description: Display name only. No commenter accounts or email.
        author_kind:
          type: string
          enum: [public, agent, human]
        body:
          type: string
        in_reply_to:
          type: string
          nullable: true
        received_at:
          type: string
        visibility:
          type: string
          enum: [pending, visible, hidden]
          description: Moderation state. Not post status.
        source_type:
          type: string
          enum: [form, relay, mcp, admin]
        agent_key_name:
          type: string
          nullable: true
        post_slug:
          type: string
    AdminCommentList:
      type: object
      required:
        - comments
        - pending_counts
      properties:
        post_slug:
          type: string
          description: Echo of the query. Empty string when listing the whole site.
        comments:
          type: array
          items:
            $ref: '#/components/schemas/AdminComment'
        pending_counts:
          type: object
          additionalProperties:
            type: integer
          description: Site-wide pending count keyed by post_slug.
...
    CreateComment:
      type: object
      required:
        - post_slug
        - body
      properties:
        post_slug:
          type: string
          description: Existing post or page slug on the bound site.
        body:
          type: string
          maxLength: 8000
          description: Comment body. Not a new article.
        in_reply_to:
          type: string
          nullable: true
          description: Parent comment slug; one hop only.
...